# api/openapi/courses-organization.yaml
openapi: 3.0.3
info:
  title: Authena course-organization service
  description: service for course managing
  version: 0.0.1
servers:
  - url: 'https://{hostname}/v1'
    variables:
      hostname:
        default: localhost
security:
  - bearerAuth: [ ]
paths:
  /courses:
    post:
      tags:
        - courses
      operationId: createCourse
      requestBody:
        description: course creation request data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCourseRequest'
      responses:
        '201':
          headers:
            Content-Location:
              description: created course url
              schema:
                type: string
          description: course created
        '400':
          description: invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: only authorized teacher can create course
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - courses
      operationId: getAllCourses
      parameters:
        - in: query
          name: title
          schema:
            type: string
          required: false
          description: course title substring for filtering
      responses:
        '200':
          description: found courses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllCoursesResponse'
        '401':
          description: authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /courses/{courseId}:
    get:
      tags:
        - courses
      operationId: getCourse
      parameters:
        - in: path
          name: courseId
          schema:
            type: string
            format: uuid
          required: true
          description: course id
      responses:
        '200':
          description: found course
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCourseResponse'
        '404':
          description: course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - courses
      operationId: editCourse
      parameters:
        - in: path
          name: courseId
          schema:
            type: string
            format: uuid
          required: true
          description: course id
      requestBody:
        description: course editing request data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditCourseRequest'
      responses:
        '204':
          description: course edited
        '400':
          description: invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: only authorized teacher can edit course
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /courses/{courseId}/auxiliary-materials:
    get:
      tags:
        - auxiliary-materials
      operationId: getAllCourseAuxiliaryMaterials
      parameters:
        - in: path
          name: courseId
          schema:
            type: string
            format: uuid
          required: true
          description: course id
        - in: query
          name: resourceType
          schema:
            $ref: '#/components/schemas/ResourceType'
          required: false
          description: resource type for filtering
      responses:
        '200':
          description: found auxiliary materials for corse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllAuxiliaryMaterialsResponse'
        '401':
          description: authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - auxiliary-materials
      operationId: attachAuxiliaryMaterialToCourse
      parameters:
        - in: path
          name: courseId
          schema:
            type: string
            format: uuid
          required: true
          description: course id
      requestBody:
        description: attaching auxiliary material request body
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddAuxiliaryMaterialRequest'
      responses:
        '201':
          description: auxiliary material attached
          headers:
            Content-Location:
              description: attached auxiliary material url
              schema:
                type: string
        '400':
          description: invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: only authorized teacher can attach auxiliary material
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /courses/{courseId}/extended:
    post:
      tags:
        - courses
      operationId: extendCourse
      parameters:
        - in: path
          name: courseId
          schema:
            type: string
            format: uuid
          required: true
          description: course id
      requestBody:
        description: course extending request data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtendCourseRequest'
      responses:
        '201':
          description: course extended
          headers:
            Content-Location:
              description: created course url
              schema:
                type: string
        '400':
          description: invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: only authorized teacher can extend course
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /courses/{courseId}/tasks/{taskNumber}:
    get:
      tags:
        - tasks
      operationId: getCourseTask
      description: returns task of course
      parameters:
        - in: path
          name: courseId
          schema:
            type: string
            format: uuid
          required: true
          description: course id
        - in: path
          name: taskNumber
          schema:
            type: integer
            minimum: 1
          required: true
          description: task number
      responses:
        '200':
          description: found task in course
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/ManualCheckingTaskResponse'
                  - $ref: '#/components/schemas/AutoCodeCheckingTaskResponse'
                  - $ref: '#/components/schemas/TestingTaskResponse'
                discriminator:
                  propertyName: type
        '404':
          description: course or task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CreateCourseRequest:
      type: object
      required: [ started, title, period ]
      properties:
        started:
          type: boolean
        title:
          type: string
        period:
          $ref: '#/components/schemas/CoursePeriod'
    GetAllCoursesResponse:
      type: array
      items:
        $ref: '#/components/schemas/Course'
    GetCourseResponse:
      $ref: '#/components/schemas/Course'
    ExtendCourseRequest:
      type: object
      required: [ started, title ]
      properties:
        started:
          type: boolean
        title:
          type: string
        period:
          $ref: '#/components/schemas/CoursePeriod'
    EditCourseRequest:
      type: object
      properties:
        started:
          type: boolean
        title:
          type: string
        period:
          $ref: '#/components/schemas/CoursePeriod'
...
    ManualCheckingTaskResponse:
      allOf:
        - $ref: '#/components/schemas/TaskResponse'
        - $ref: '#/components/schemas/ManualCheckingTaskPart'
    AutoCodeCheckingTaskResponse:
      allOf:
        - $ref: '#/components/schemas/TaskResponse'
        - $ref: '#/components/schemas/AutoCodeCheckingTaskPart'
    TestingTaskResponse:
      allOf:
        - $ref: '#/components/schemas/TaskResponse'
        - $ref: '#/components/schemas/TestingTaskPart'
...
    GetAllAuxiliaryMaterialsResponse:
      type: array
      items:
        $ref: '#/components/schemas/AuxiliaryMaterial'
    AddAuxiliaryMaterialRequest:
      $ref: '#/components/schemas/AuxiliaryMaterial'
...
    Course:
      type: object
      required: [ id, title, period, creatorId, started, tasksNumber ]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        period:
          $ref: '#/components/schemas/CoursePeriod'
        creatorId:
          type: string
          format: uuid
        started:
          type: boolean
        tasksNumber:
          type: integer
          minimum: 0
    CoursePeriod:
      type: object
      required: [ academicStartYear, academicEndYear, semester ]
      properties:
        academicStartYear:
          type: integer
        academicEndYear:
          type: integer
        semester:
          $ref: '#/components/schemas/Semester'
...
    ResourceType:
      type: string
      enum:
        - TRAINING_MANUAL
        - VIDEO
        - PRESENTATION
        - OTHER
    Error:
      type: object
      required: [ slug, details ]
      properties:
        slug:
          type: string
        details:
          type: string